+++ src/components/header/Alert.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Alert = () => {
+    return (
+        <div>
+            
+        </div>
+    );
+};
+
+export default Alert;
